RemoveUserSubNodes CLI migration
